requirements.txt
absl-py==0.9.0
blessings==1.7
certifi==2019.11.28
dominate==2.4.0
grpcio==1.16.1
Markdown==3.1.1
numpy==1.18.1
nvidia-ml-py3==7.352.0
olefile==0.46
opencv-python==4.2.0.32
Pillow==7.0.0
protobuf==3.11.3
psutil==5.7.0
pycocotools
scikit-image
scipy==1.4.1
six==1.14.0
tensorboard==2.0.0
tensorboardX==2.0
torch==1.4.0
torchprofile==0.0.1
torchvision==0.5.0
tqdm==4.42.1
Werkzeug==1.0.0
wget==3.2